requirements.txt
streamlit
requests
PyPDF2
docx2txt
python-dotenv
google.generativeai
sqlalchemy